Remove canonical URL
I don't think we should be defining the URL in the site itself if we can help it.
As much as possible, the static site should be self-contained and host-agnostic.

There's every chance, in the event that the domain changes, that we will forget
to update it here in the codebase, as nothing will break in any obvious way.

To the best of our ability, we should try to make sure the site is publicly
available at only one location anyway, rendering "rel=canonical" redundant. If we
do have a clear business case for "rel=canonical" we should add it at that stage.